refactor: keep request context internal
