+++ lib/posthog.rb
@@ -10,7 +10,6 @@ require 'posthog/transport'
 require 'posthog/response'
 require 'posthog/logging'
 require 'posthog/exception_capture'
-require 'posthog/context'
 require 'posthog/feature_flag_error'
 require 'posthog/feature_flag_result'
 require 'posthog/feature_flag_evaluations'
